requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
numpy>=1.24.0
python-dotenv>=1.0.0 